move: 모달 상태 관리 path 이동
